requirements.txt
playwright>=1.44.0
fastapi>=0.111.0
uvicorn>=0.30.0
anthropic>=0.28.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.7.0
